requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
mysql-connector==2.1.6